Add performance lever to pull; make checking for city optional.
Introduces a new option check_city_exists. If set to True, then recata
will check if the city exists before fetching the city.id for inserting
the listing. If set to False, recata will assume the city exists and try
to fetch the city id without checking if it exists first.

When the cities table is prepopulated, this option makes sense to use.